Add the model files, make changes to the parameters in the deepface find function and to the code that marks the attendance
Made the model to generate its required files and also set silent parameter to True to suppress the warning and logs that were being displayed on the terminal. Also made changes to the code that marked the attendance. Now the script will mark the attendance only when the user presses Enter.